♻️ Moved APIs to routes
